Use INT32_MAX as max row size of segment file& change v2 num_rows type from uint64_t to uint32_t

+++ be/src/olap/rowset/segment_v2/segment_writer.cpp
@@ -101,7 +101,7 @@ uint64_t SegmentWriter::estimate_segment_size() {
     return 0;
 }
 
-uint64_t SegmentWriter::num_rows_written() {
+uint32_t SegmentWriter::num_rows_written() {
     return _row_count;
 }
